refactor(logical-plan): split LogicalPlanner into single-responsibility phases (#851 PR2a)
Split the 1,576-line logical-planner.ts god object along its existing seams.
LogicalPlanner is now a thin facade composing four phase classes plus shared utils:

- JoinPlanner            (join-planner.ts)        — join-plan construction
- CTEPlanner             (cte-planner.ts)         — pre-aggregation CTE / fan-out logic
- FilterPropagation      (filter-propagation.ts)  — filter propagation into CTEs
- PlanAnalysisReporter   (plan-analysis-reporter.ts) — dry-run/EXPLAIN trace + warnings
- ResolverCache + analyzeCubeUsage/extractCubeNamesFromFilter (planner-utils.ts)

Dedup: removed the unreachable findJoinInfoToCube() (only called behind an
`isPrimary` flag hard-coded to false; subsumed by findJoinInfoForCube) and the
now-dead isPrimary branches.

LogicalPlanner's public surface (analyzeCubeUsage/analyzePrimaryCube/
analyzeJoinPathForTarget/buildJoinPlanForPrimary/buildPreAggregationCTEs/
buildWarnings) is preserved as facade delegations — used by LogicalPlanBuilder
and the planner unit tests.

Behaviour-preserving: typecheck clean; 2348 server tests pass (postgres).
